added the input box for free response

+++ app/controllers/user_responses_controller.rb
@@ -1,25 +1,42 @@
 class UserResponsesController < ApplicationController
   def create
-    @user_response = UserResponse.new(
-      response: Response.find(params[:response_id]),
-      user: current_user
-    )
-    @response = @user_response.response
-    @user_response.save
-    respond_to do |format|
-      format.js { render :insert }
-      format.html { redirect_to questions_path }
+    unless params[:response_id] == 194
+      @user_response = UserResponse.new(
+        response: Response.find(params[:response_id]),
+        user: current_user
+      )
+      @response = @user_response.response
+      @user_response.save
+      respond_to do |format|
+        format.js { render :insert }
+        format.html { redirect_to questions_path }
+      end
+    else
+
+      @user_response = UserResponse.new(
+        response: Response.find(params[:response_id]),
+        user: current_user,
+        free_response: params[:user_response][:free_response]
+      )
+      @user_response.save
+      redirect_to questions_path
     end
   end
 
   def destroy
-    user_response = UserResponse.find(params[:id])
-    @response = user_response.response
-    user_response.destroy
-    @user_response = nil
-    respond_to do |format|
-      format.js { render :insert }
-      format.html { redirect_to questions_path }
+    unless params[:response_id] == 194
+      user_response = UserResponse.find(params[:id])
+      @response = user_response.response
+      user_response.destroy
+      @user_response = nil
+      respond_to do |format|
+        format.js { render :insert }
+        format.html { redirect_to questions_path }
+      end
+    else
+      @user_response = UserResponse.find(params[:id])
+      @user_response.destroy
+      @user_response.save
     end
   end
 end
